requirements.txt
# Local Choose Your Own Adventure Game Dependencies
# Supports both Ollama (local LLM) and OpenAI

# Core LangChain components
langchain>=0.1.0
langchain-core>=0.1.0

# AI Provider integrations
langchain-ollama>=0.1.0       # For Ollama integration
langchain-openai>=0.1.0       # For OpenAI integration
openai>=1.0.0                 # OpenAI SDK

# Environment configuration
python-dotenv>=1.0.0          # Load environment variables

# Qdrant vector database
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0

# Local embeddings
sentence-transformers>=2.2.2

# HTTP requests for health checks
requests>=2.31.0

# Additional dependencies for the updated code
pydantic>=2.0.0

# Optional: For better performance
torch>=2.0.0  # Only if you want GPU acceleration for embeddings
